randomize all done
